ai-service/requirements.txt
Flask==2.2.2
flask-cors==4.0.0
google-generativeai==0.3.2
openai-whisper
openai==1.3.7
gunicorn==20.1.0
Werkzeug==2.2.2
python-dotenv==1.0.0
pydub==0.25.1
ffmpeg-python==0.2.0
numpy>=2.0,<2.3
genanki==0.13.1